article/code/requirements.txt
attrs==22.1.0
biopython==1.79
blis==0.7.7
brotlipy==0.7.0
catalogue==2.0.7
charset-normalizer==2.0.12
click==7.1.2
coloredlogs==15.0.1
conllu==4.5.2
cssselect==1.1.0
cymem==2.0.6
distlib==0.3.5
docopt==0.6.2
en-tox @ https://huprdontoxnlpst.blob.core.windows.net/en-tox/en_tox-0.1.0-py3-none-any.whl
eutils==0.6.0
filelock==3.8.0
habanero==1.2.2
humanfriendly==10.0
iniconfig==1.1.1
Jinja2==3.1.2
joblib==1.1.0
lxml==4.9.1
MarkupSafe==2.1.1
metapub==0.5.5
murmurhash==1.0.7
nmslib==2.1.1
numpy 
packaging==21.3
pandas==1.3.4
pathy==0.6.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.6
psutil==5.9.1
py==1.11.0
pybind11==2.6.1
pydantic==1.8.2
pyparsing==3.0.9
pysbd==0.3.4
pytest==7.1.2
python-Levenshtein==0.12.2
pytz==2022.1
PyYAML==6.0
regex==2022.9.13
scikit-learn==1.1.2
scipy==1.8.1
scispacy==0.4.0
smart-open==5.2.1
spacy==3.0.8
spacy-legacy==3.0.9
srsly==2.4.3
tabulate==0.8.10
thinc==8.0.16
threadpoolctl==3.1.0
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tox==3.25.1
tqdm==4.64.0
transformers==4.23.1
typer==0.3.2
Unidecode==1.3.4
virtualenv==20.16.3
wasabi==0.9.1
wcwidth
